refactor: 기존 통합 테스트를 단위 테스트로 완전히 변경
